pyproject.toml
[project]
name = "bio-code-interpreter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.72.1",
    "e2b-code-interpreter>=2.3.0",
    "fastapi>=0.121.1",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
